Support applying the plugin to any project in hierarchy
This change removes accesses to the extension via `rootProject`.
Instead, we now access the extension wherever it was actually created.

This means that when applying the plugin to a project that is not the rootProject, we will use the correct extension instead of looking for it in the rootProject (where it is not).

Resolves:
https://github.com/Kotlin/binary-compatibility-validator/issues/37